add main.rs and crate separation
