+++ .dockerignore
@@ -1,6 +1,4 @@
 *
-!build.rs
 !Cargo.toml
 !Cargo.lock
 !src
-!config
