+++ Cargo.toml
@@ -26,6 +26,7 @@ thiserror = "2"
 prometheus-client = "0.22"
 tracing-subscriber = "0.3"
 tower-http = { version = "0.6", features = ["trace"] }
+clap = { version = "4.5.23", features = ["derive", "env"] }
 
 [dev-dependencies]
 serde_test = "1"
